Rename ibeis import references to wbia
We are in the middle of renaming our repos from ibeis to wbia.

The changes are from running this command:

```
git grep -l ibeis utool | xargs sed -i \
    -e 's/\(from\|import\) ibeis\(\.\| \|$\)/\1 wbia\2/g' \
    -e 's/\(from\|import\) guitool_ibeis/\1 ibeis.guitool/g' \
    -e 's/\([^_]\)ibeis\./\1wbia./g'
```

and manually changing 2 instances of `guitool_ibeis` to `wbia.guitool`
in utool/Preferences.py

It seems we forgot to rename `guitool_ibeis` to `ibeis.guitool`, so just
skip that step and rename `guitool_ibeis` to `wbia.guitool`.