Use find_packages() in setup.py to fix missing module in wheel
When installing `wbia-utool` using the wheel on pypi, `utool.experimental` is
not available:

```
  File "/home/runner/work/wildbook-ia/wildbook-ia/wbia/algo/verif/vsone.py", line 925, in task_evaluation_report
    from utool.experimental.pandas_highlight import to_string_monkey
ModuleNotFoundError: No module named 'utool.experimental'
```

The directory `experimental` isn't in the wheel.  The reason seems to be that
we manually set the packages to a specific list which didn't include
`utool.experimental`.  To avoid having the same error, just use
`setuptools.find_packages`.